Move from Agent wrapper to pure data abstraction.
The Agent wrapper was in hindsight completely unnecessary. All the
functions now work on the defined Structs themselves and if you need
to retain state over process - then you are free to use an Agent
whose initial state is that of the `TableRex.Table` struct.

Doing this allows much greater freedom as to where this library can
be used. The table now exists in known structs and is only modified
by using the applicable functions.