Add Table.new/3
`Table.new/0` has been supplemented with `Table.new/3` which takes `rows`
and an optional `header` and `title`. This change was made as when the data
is known upfront and the user wanted a custom table, it was quite verbose
calling `Table.new/3` and then calling `add_rows`, `set_header` and
`set_title`.

See CHANGELOG for more information and an example.