Table.set_* -> Table.put_*
This change is being made to better fit in with Elixir convention;
"set" implies the existence of state which is not correct, "put" is
used over "update" as none of the functions use the current value
and are simply overriding functions. This usage is inline with say
the Dict module where put and update both exist.