Convert sort/2 to sort/3 before release (#29)
sort/2 was brought in by #27 but has yet to be released, because
of that I took the decision to make it support sort direction
(ascending/descending) before release, which it now does - see
docs for usage.

As part of this, a breaking change has been made to the internal
data structure which represents a Cell: they now store the original
data type _and_ the rendered to string version, which allows us to
do some more advanced functions in the future (including better
custom sorting). The stringified value of a Cell used to be stored
at `value`, it is now stored at `rendered_value` - with the original
being stored at `raw_value`. This is technically not a breaking change
to a public API as the Cell's are largely hidden by the implementation
_but_ I will still increment the major version so as to not catch anyone
unaware.