Implement multiline cell rendering (#39)
Add multiline rendering support to the text renderer as a
backwards-compatible upgrade. Previously the text renderer
could only support rendering cells with a single line of
text with no newline breaks, and adding newline breaks
caused the display of the entire table to break.

The new feature has the same API as before, simply pass in
a string containing newline chars/line feeds instead of the
usual cell input strings and it will be handled correctly
including coloring and alignment.

Thanks very much to @rubysolo for this feature!